Moving app contents to a HeliumPlace component, moving Map-specific legends to Map component
